+++ src/Benzene.Aws.Lambda.Sns/DependencyInjectionExtensions.cs
@@ -4,6 +4,7 @@ using Benzene.Abstractions.MessageHandlers.Mappers;
 using Benzene.Abstractions.MessageHandlers.Request;
 using Benzene.Abstractions.Messages.Mappers;
 using Benzene.Core.MessageHandlers.Info;
+using Benzene.Core.MessageHandlers.MediaFormats;
 using Benzene.Core.MessageHandlers.Request;
 using Benzene.Core.MessageHandlers.Serialization;
 
@@ -31,9 +32,10 @@ public static class DependencyInjectionExtensions
         services.AddScoped<IMessageHeadersGetter<SnsRecordContext>, SnsMessageHeadersGetter>();
         services.AddScoped<IMessageBodyGetter<SnsRecordContext>, SnsMessageBodyGetter>();
         services.AddScoped<IMessageHandlerResultSetter<SnsRecordContext>, SnsMessageMessageHandlerResultSetter>();
+        services.AddMediaFormatNegotiation<SnsRecordContext>();
         services
             .AddScoped<IRequestMapper<SnsRecordContext>,
-                MultiSerializerOptionsRequestMapper<SnsRecordContext, JsonSerializer>>();
+                MultiSerializerOptionsRequestMapper<SnsRecordContext>>();
 
         services.AddSingleton<ITransportInfo>(_ => new TransportInfo("sns"));